Перенос работы с симптомами болезни в модуль болезней персонажа
