+++ Zilon.Core/Zilon.Core/Persons/DiseaseData.cs
@@ -1,7 +1,9 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
 using System.Linq;
 
 using Zilon.Core.Diseases;
+using Zilon.Core.Persons.Survival;
 
 namespace Zilon.Core.Persons
 {
@@ -37,6 +39,89 @@ namespace Zilon.Core.Persons
         {
             var currentProcess = _diseases.SingleOrDefault(x => x.Disease == disease);
             _diseases.Remove(currentProcess);
+        }
+
+        public void Update(EffectCollection personEffects)
+        {
+            foreach (var diseaseProcess in Diseases.ToArray())
+            {
+                // Если есть болезнь, то назначаем эффекты на симпомы этой болезни.
+
+                // Первые 25% силы болезнь себя не проявляет.
+                // Оставшиеся 75% делим на равные отрезки между всеми симптомами.
+                // По мере увеличения силы болезни добавляем по эффекту на каждый симптом за каждый пройденный
+                // рассчитанный отрезок на симптом.
+                // Если эффект на такой симптом уже есть, то добавляем к этому симптому новую болезнь.
+                // После 50% прогресса болезни (то есть когда прошёл пик силы) начинаем обратный процесс - убираем эффекты
+                // за каждый симптом.
+                // Если для эффекта симптома указано несколько болезней, то эффект уходит только после удаления последней болезни.
+
+                var disease = diseaseProcess.Disease;
+                var symptoms = disease.GetSymptoms();
+
+                var currentPower = diseaseProcess.CurrentPower;
+
+                // Рассчитываем отрезок силы на один симптом.
+                var symptomPowerSegment = 0.75f / symptoms.Length;
+
+                if (diseaseProcess.Value < 0.5f)
+                {
+                    // Обрабатываем нарастание силы болезни.
+
+                    if (currentPower > 0.25f)
+                    {
+                        // Симптомы начинаю проявляться после 25% силы болезни.
+
+                        // Рассчитываем количество симптомов, которые должны быть при текущей силе болезни.
+
+                        var activeSymptomCount = (int)Math.Ceiling((currentPower - 0.25f) / symptomPowerSegment);
+
+                        // Начинаем проверять, есть ли эффекты на все эти симптомы
+                        // или добавлены ли болезни симптомов в список болезней эффектов.
+
+                        for (var i = 0; i < activeSymptomCount; i++)
+                        {
+                            var currentSymptom = symptoms[i];
+
+                            var currentSymptomEffect = personEffects.Items.OfType<DiseaseSymptomEffect>()
+                                .SingleOrDefault(x => x.Symptom == currentSymptom);
+
+                            if (currentSymptomEffect is null)
+                            {
+                                currentSymptomEffect = new DiseaseSymptomEffect(disease, currentSymptom);
+                                personEffects.Add(currentSymptomEffect);
+                            }
+                            else
+                            {
+                                currentSymptomEffect.AddDisease(disease);
+                            }
+                        }
+                    }
+                }
+                else
+                {
+                    var activeSymptomCount = (int)Math.Ceiling(currentPower / symptomPowerSegment);
+
+                    // Начинаем снимать все эффекты, которые за пределами количества.
+
+                    var symptomLowerIndex = activeSymptomCount - 1;
+
+                    for (var i = symptomLowerIndex; i < symptoms.Length; i++)
+                    {
+                        var currentSymptom = symptoms[i];
+
+                        var currentSymptomEffect = personEffects.Items.OfType<DiseaseSymptomEffect>()
+                            .Single(x => x.Symptom == currentSymptom);
+
+                        currentSymptomEffect.RemoveDisease(disease);
+
+                        if (!currentSymptomEffect.Diseases.Any())
+                        {
+                            personEffects.Remove(currentSymptomEffect);
+                        }
+                    }
+                }
+            }
         }
     }
 }
