+++ Zilon.Core/Zilon.Core/Persons/IDiseaseData.cs
@@ -25,5 +25,11 @@ namespace Zilon.Core.Persons
         /// </summary>
         /// <param name="disease"> Целевая болезнь. </param>
         void RemoveDisease(IDisease disease);
+
+        /// <summary>
+        /// Обновление состояния модуля болезней.
+        /// </summary>
+        /// <param name="personEffects"> Ссылка на модуль эффектов персонажа. Болезни навешивают эффекты. </param>
+        void Update(EffectCollection personEffects);
     }
 }
